requirements.txt
starlette==0.27.0
uvicorn==0.20.0
selenium==4.11.2
webdriver-manager==3.8.6
python-dotenv==1.0.0